Forecast a model from the fable ATA model

Forecast a model from the fable ATA model

# S3 method for ATA
forecast(
  object,
  new_data,
  h = NULL,
  ci_level = 95,
  negative_forecast = TRUE,
  onestep = FALSE,
  ...
)

Arguments

object

The time series model used to produce the forecasts

new_data

A `tsibble` containing future information used to forecast.

h

The forecast horison (can be used instead of `new_data` for regular time series with no exogenous regressors).

ci_level

Confidence Interval levels for forecasting. Default value is 95.

negative_forecast

Negative values are allowed for forecasting. Default value is TRUE. If FALSE, all negative values for forecasting are set to 0.

onestep

Default is FALSE. if TRUE, the dynamic forecast strategy uses a one-step model multiple times `h` forecast horizon) where the prediction for the prior time step is used as an input for making a prediction on the following time step.

...

Other arguments

Value

A vector of fitted residuals.

Examples

library(fable.ata)
as_tsibble(USAccDeaths) %>%
  model(ata = AutoATA(value ~ trend("A") + season("M"))) %>% forecast(h=24)
#> # A fable: 24 x 4 [1M]
#> # Key:     .model [1]
#>    .model    index    value  .mean
#>    <chr>     <mth>   <dist>  <dbl>
#>  1 ata    1979 Jan 8091.851  8092.
#>  2 ata    1979 Feb 7352.286  7352.
#>  3 ata    1979 Mar 8154.435  8154.
#>  4 ata    1979 Apr 8387.586  8388.
#>  5 ata    1979 May 9272.331  9272.
#>  6 ata    1979 Jun 9691.667  9692.
#>  7 ata    1979 Jul 10632.91 10633.
#>  8 ata    1979 Aug 9921.513  9922.
#>  9 ata    1979 Sep  8799.18  8799.
#> 10 ata    1979 Oct 9183.734  9184.
#> # ℹ 14 more rows